Simplify reader interface to only 1 flag cache_file for this behaivor:
For LocalReader:

If cache_files is True, it caches the file paths.
If False, it retrieves the file list each time.

For S3Reader (and other cloud readers):

If cache_files is True, it downloads files to a local directory and reads from there.
If False, it reads directly from remote storage each time.